calculateacceleration

+++ Entropic Controller/A.cpp
@@ -88,7 +88,7 @@ long long int calculateacceleration(particle particle, short dimension) {
 	//Calculate the force on this particle by summing the force on all other particles in the system
     double force = 0;
 	//Loop through all particles in the system
-    for (int i = 0; i < particles.size(); i++) {
+    for (unsigned long long int i = 0; i < particles.size(); i++) {
         if (i != particle.number) { //If the particle is the same as the input particle, we skip it
             //Calculate the distance between the two particles
             distance = 2E-19 * abs(particle.position[dimension] - particles[i].position[dimension]); //Multiply the calculated unit distance by 0.2 attometers
